Move OperationType to scylla-cql-core
