+++ scylla/src/errors.rs
@@ -35,8 +35,9 @@ pub use scylla_cql::frame::frame_errors::{
     FrameHeaderParseError,
 };
 pub use scylla_cql::frame::response::CqlResponseKind;
-pub use scylla_cql::frame::response::error::{DbError, OperationType, WriteType};
+pub use scylla_cql::frame::response::error::{DbError, WriteType};
 pub use scylla_cql_core::frame::request::CqlRequestKind;
+pub use scylla_cql_core::frame::response::error::OperationType;
 pub use scylla_cql_core::serialize::SerializationError;
 
 /// Error that occurred during request execution
